+++ src/decoder/pilgrim2text.c
