+++ src/decoder/pilgrim_cfg_decoder.c
@@ -3,52 +3,24 @@
  *     See COPYRIGHT in top-level directory
  */
 
-#include <mpi.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <assert.h>
 #include "pilgrim.h"
 #include "pilgrim_reader.h"
+#include "pilgrim_utils.h"
 #include "uthash.h"
-#include "dlmalloc-2.8.6.h"
 
+/**
+ * A hash table to represent a CFG
+ * Key: rule id
+ * Val: rule body
+ */
 typedef struct RuleHash_t {
     int rule_id;
     int *rule_body;     // 2i+0: val of symbol i,  2i+1: exp of symbol i
-    int symbols;        // how many symbols in the rule body
+    int symbols;        // There are a total of 2*symbols integers in the rule body
     UT_hash_handle hh;
 } RuleHash;
-
-typedef struct VariablePool_t {
-    int value;
-    int id;
-    UT_hash_handle hh;
-} VariablePool;
-
-static int rank;
-static int nprocs;
-static RuleHash* rules_table;
-static VariablePool* variable_pools[15];    // One variable pool for each variable type
-static int variable_current_ids[15];
-
-void read_global_metadata(char* path, GlobalMetadata *gm) {
-    char global_metadata_path[256];
-    sprintf(global_metadata_path, "%s/pilgrim.mt", path);
-    FILE* fh = fopen(global_metadata_path, "rb");
-    fread(gm, sizeof(GlobalMetadata), 1, fh);
-    fclose(fh);
-
-    printf("Total procs: %d, Time resolution: %.1fus\n\n", gm->ranks, gm->time_resolution*1000000);
-}
-
-void read_local_metadata(char* path, int rank, LocalMetadata *lm) {
-    char local_metadata_path[256];
-    sprintf(local_metadata_path, "%s/%d.mt", path, rank);
-    FILE* fh = fopen(local_metadata_path, "rb");
-    fread(lm, sizeof(LocalMetadata), 1, fh);
-    fclose(fh);
-    printf("rank: %d, tstart: %f, tend: %f, records: %ld\n", lm->rank, lm->tstart, lm->tend, lm->records_count);
-}
 
 void print_rule(RuleHash *rule) {
     printf("rule %d, symbols: %d\n\t-->", rule->rule_id, rule->symbols);
@@ -58,87 +30,41 @@ void print_rule(RuleHash *rule) {
 }
 
 
-/*
- * First pass: decompress the inter-process compressed grammars
+/**
+ * Decompress a CFG by recursive rule application
  *
- * Recursively decode rules
+ * The decompressed symbols will be represented by
+ * decopmressed_symbols array.
+ *
  */
-int decode_rule(int* decompressed_symbols, int rule_id, int start_rule_id) {
-    int advance = 0;
-
+void rule_application(RuleHash* rules, int rule_id, int start_rule_id, int* decoded_symbols, int* pos) {
     RuleHash *rule;
-    HASH_FIND_INT(rules_table, &rule_id, rule);
+    HASH_FIND_INT(rules, &rule_id, rule);
 
     for(int i = 0; i < rule->symbols; i++) {
         int sym_val = rule->rule_body[2*i+0];
         int sym_exp = rule->rule_body[2*i+1];
         // Non-terminal, i.e., a rule
         if(sym_val < start_rule_id) {
-            for(int j = 0; j < sym_exp; j++) {
-                int advanced = decode_rule(decompressed_symbols, sym_val, start_rule_id);
-                decompressed_symbols += advanced;
-                advance += advanced;
-            }
+            for(int j = 0; j < sym_exp; j++)
+                rule_application(rules, sym_val, start_rule_id, decoded_symbols, pos);
         }
         // Terminal
         else {
-            *decompressed_symbols = sym_val;
-            decompressed_symbols++;
-            *decompressed_symbols = sym_exp;
-            decompressed_symbols++;
-            advance += 2;
+            if(decoded_symbols) {
+                decoded_symbols[*pos]= sym_val;
+                decoded_symbols[*pos+1] = sym_exp;
+            }
+            *pos = *pos + 2;
         }
     }
-
-    return advance;
 }
-
-char* get_variable_name(CallSignature *cs, int i) {
-    int type = cs->arg_types[i];
-    int direction = cs->arg_directions[i];
-
-    char *name = calloc(sizeof(char), 20);
-    if(type == TYPE_NON_MPI) {
-        sprintf(name, "[Not Handled]");
-    } else if(type == TYPE_MPI_Comm) {
-        sprintf(name, "%s_0", TYPE_VAR_STR[type]);
-    } else {    // all other mpi objects with integer values
-        int value = *((int*)cs->args[i]);
-        VariablePool* var = NULL;
-        HASH_FIND_INT(variable_pools[type], &value, var);
-
-        assert(var);
-
-        if(direction == DIRECTION_IN)
-            sprintf(name, "%s_%d", TYPE_VAR_STR[type], var->id);
-        else
-            sprintf(name, "&%s_%d", TYPE_VAR_STR[type], var->id);
-    }
-
-    return name;
-}
-
-
-void write_call(FILE* f, CallSignature *cs) {
-    fprintf(f, "%s(", func_names[cs->func_id]);
-    for(int i = 0; i < cs->arg_count; i++) {
-        char* var = get_variable_name(cs, i);
-        fprintf(f, "%s", var);
-        free(var);
-
-        if(i != cs->arg_count - 1)
-            fprintf(f, ", ");
-        else
-            fprintf(f, ");\n");
-    }
-}
-
 
 /**
  * Second pass: for each rank, decode its grammar
  */
-void decode_and_write(int rule_id, FILE *f, CallSignature *call_sigs) {
-
+/*
+void decode_and_write(int rule_id, FILE *f) {
     RuleHash *rule;
     HASH_FIND_INT(rules_table, &rule_id, rule);
     for(int i = 0; i < rule->symbols; i++) {
@@ -147,18 +73,21 @@ void decode_and_write(int rule_id, FILE *f, CallSignature *call_sigs) {
 
         // Non-terminal, i.e., a rule
         if(sym < -1) {
-            decode_and_write(sym, f, call_sigs);
+            decode_and_write(sym, f);
         }
         // Terminal
         else {
-            CallSignature *cs = &(call_sigs[sym]);
-            write_call(f, cs);
+            if(sym == 9999999)
+                fprintf(f, "0\n");
+            else
+                fprintf(f, "%.6f\n", pow(1.1, sym));
         }
     }
 }
+*/
 
 
-void clean_rules_table() {
+void clean_rules(RuleHash* rules_table) {
     RuleHash *current, *tmp;
     HASH_ITER(hh, rules_table, current, tmp) {
         HASH_DEL(rules_table, current);
@@ -168,21 +97,25 @@ void clean_rules_table() {
     rules_table = NULL;
 }
 
-void read_cfg(char *path, int total_ranks, CallSignature* funcs) {
-    printf("\nRead CFG\n");
-    char grammar_file_path[256];
-    sprintf(grammar_file_path, "%s/grammars.dat", path);
+/**
+ * Read the inter-process compressed grammar
+ * And store its rules in a hash-table.
+ *
+ * It will be later used for decopmression.
+ */
+RuleHash* read_inter_compressed_grammar(char* path, size_t* original_integers, int* start_rule_id) {
 
-    FILE* f = fopen(grammar_file_path, "rb");
+    FILE* f = fopen(path, "rb");
 
-    int start_rule_id, rules;
-    size_t uncompressed_integers;
-    fread(&start_rule_id, sizeof(int), 1, f);
-    fread(&uncompressed_integers, sizeof(size_t), 1, f);
+    // Read the inter-process compressed grammar
+    // The header has three values:
+    // start rule id, number of rules, and number of intergers once decompressed.
+    int rules;
+    fread(start_rule_id, sizeof(int), 1, f);
+    fread(original_integers, sizeof(size_t), 1, f);
     fread(&rules, sizeof(int), 1, f);
 
-    printf("Start_rule_id: %d, Rules: %d, Uncompressed integers: %ld\n", start_rule_id, rules, uncompressed_integers);
-
+    RuleHash* rules_table = NULL;
     for(int i = 0; i < rules; i++) {
         RuleHash *rule = malloc(sizeof(RuleHash));
 
@@ -192,168 +125,83 @@ void read_cfg(char *path, int total_ranks, CallSignature* funcs) {
         rule->rule_body = (int*) malloc(sizeof(int)*rule->symbols*2);
         fread(rule->rule_body, sizeof(int), rule->symbols*2, f);
 
-        print_rule(rule);
+        HASH_ADD_INT(rules_table, rule_id, rule);
+    }
+    printf("Start_rule_id: %d, Rules: %d, Uncompressed integers: %ld\n", *start_rule_id, rules, *original_integers);
+
+    return rules_table;
+}
+
+/**
+ * After decopmress the inter-process compressed grammar into
+ * an array, we can read every ranks copmressed grammar.
+ *
+ * This function will read the grammar (integer array) to form
+ * a hash table of rules
+ */
+RuleHash* read_intra_compressed_grammar(int* grammar, int *size) {
+    int pos = 0;
+    RuleHash* rules_table = NULL;
+    int rules = grammar[pos++];
+    pos++;
+    for(int i = 0; i < rules; i++) {
+        RuleHash *rule = malloc(sizeof(RuleHash));
+
+        rule->rule_id = grammar[pos++];
+        pos++;  // skip exp
+
+        rule->symbols = grammar[pos++];
+        pos++;  // skip exp
+
+        rule->rule_body = malloc(2*sizeof(int)*rule->symbols);
+        memcpy(rule->rule_body, &(grammar[pos]), 2*sizeof(int)*rule->symbols);
+        pos += (rule->symbols)*2;
         HASH_ADD_INT(rules_table, rule_id, rule);
     }
 
-    int *decompressed  = malloc(sizeof(int) * uncompressed_integers);
-    decode_rule(decompressed, start_rule_id, start_rule_id);
-    clean_rules_table();
+    *size = pos;
+    return rules_table;
+}
 
+
+/**
+ * Decompress the inter-process compressed grammar
+ * The grammar is read by read_compressed_grammar()
+ *
+ */
+int** read_cfg(char* path, int nprocs, int *num_symbols) {
+
+    // 1. Read and decompress inter-process compressed grammar
+    size_t original_integers;
+    int start_rule_id;
     int pos = 0;
-    for(int rank = 0; rank < total_ranks; rank++) {
+    RuleHash* inter_cfg = read_inter_compressed_grammar(path, &original_integers, &start_rule_id);
+    int* decompressed = malloc(sizeof(int)*original_integers);
+    rule_application(inter_cfg, start_rule_id, start_rule_id, decompressed, &pos);
+    clean_rules(inter_cfg);
 
-        int rules = decompressed[pos++];
-        pos++;      // skip exp
+    // 2. Read and decompress each rank's grammar
+    int** decoded_symbols = malloc(sizeof(int*)*nprocs);
 
-        for(int j = 0; j < rules; j++) {
-            RuleHash *rule = malloc(sizeof(RuleHash));
+    int last_grammar_size = 0;
+    int *grammar = decompressed;
+    for(int rank = 0; rank < nprocs; rank++) {
+        grammar += last_grammar_size;
+        RuleHash* intra_cfg = read_intra_compressed_grammar(grammar, &last_grammar_size);
 
-            rule->rule_id = decompressed[pos++];
-            pos++;  // skip exp
+        // First pass to copmute the number of symbols after decompression
+        num_symbols[rank] = 0;
+        rule_application(intra_cfg, -1, -1, NULL, &num_symbols[rank]);
 
-            rule->symbols = decompressed[pos++];
-            pos++;  // skip exp
-            printf("rank: %d, add rule: %d, symbols: %d\n", rank, rule->rule_id, rule->symbols);
+        // Second pass to fill in the symbols
+        decoded_symbols[rank] = malloc(sizeof(int) * num_symbols[rank]);
+        num_symbols[rank] = 0;
+        rule_application(intra_cfg, -1, -1, decoded_symbols[rank], &num_symbols[rank]);
 
-            rule->rule_body = malloc(sizeof(int)*rule->symbols*2);
-            rule->rule_body = malloc(2*sizeof(int)*rule->symbols);
-            memcpy(rule->rule_body, &(decompressed[pos]), 2*sizeof(int)*rule->symbols);
-            pos += (rule->symbols)*2;
-            HASH_ADD_INT(rules_table, rule_id, rule);
-        }
-
-        // write out to this rank
-        /*
-        char output_path[256];
-        sprintf(output_path, "%s/main.c", path);
-        FILE *fout = fopen(output_path, "a");
-        fprintf(fout, "\n\n//======================================\n");
-        fprintf(fout, "// Code for Rank = %d\n", rank);
-        fprintf(fout, "//======================================\n");
-        //decode_and_write(-1, fout, funcs);
-        fprintf(fout, "//======================================\n");
-        fclose(fout);
-        */
-
-        clean_rules_table();
+        clean_rules(intra_cfg);
     }
 
-    fclose(f);
     free(decompressed);
+    return decoded_symbols;
 }
 
-CallSignature* read_cst(char *directory, int *num_funcs) {
-    printf("\nRead CST\n");
-    char path[256];
-    sprintf(path, "%s/funcs.dat", directory);
-    FILE* f = fopen(path, "rb");
-
-    short func_id;
-    int entries, key_len, terminal, rank, duration, interval;
-    fread(&entries, sizeof(int), 1, f);
-    *num_funcs = entries;
-
-    CallSignature *call_sigs = malloc(sizeof(CallSignature) * entries);
-
-    char buff[512];
-    for(int i = 0; i < entries; i++) {
-
-        fread(&terminal, sizeof(int), 1, f);
-        fread(&rank, sizeof(int), 1, f);
-        fread(&key_len, sizeof(int), 1, f);
-
-        fread(&func_id, sizeof(short), 1, f);
-        fread(buff, 1, key_len-sizeof(short), f);
-
-        assert(terminal < entries);
-        call_sigs[terminal].func_id = func_id;
-        read_record_args(func_id, buff, &(call_sigs[terminal]));
-        printf("Terminal: %d, Key len: %d, Func: %s\n", terminal, key_len, func_names[func_id]);
-    }
-
-
-    fclose(f);
-    return call_sigs;
-}
-
-
-void write_init_variables(const char* path, CallSignature *call_sigs, int num_sigs) {
-
-    char output_path[256] = {0};
-    sprintf(output_path, "%s/main.c", path);
-    FILE *fout = fopen(output_path, "w");
-
-    bool comm_flag = false;
-
-    CallSignature *cs;
-    for(int i = 0; i < num_sigs; i++) {
-        cs = &(call_sigs[i]);
-        for(int j = 0; j < cs->arg_count; j++) {
-            int type = cs->arg_types[j];
-            int dir = cs->arg_directions[j];
-            if(type == TYPE_NON_MPI) {
-
-            } else if(type == TYPE_MPI_Comm) {
-                if(!comm_flag) {
-                    fprintf(fout, "%s %s_0 = MPI_COMM_WORLD;\n", TYPE_STR[type], TYPE_VAR_STR[type]);
-                    comm_flag = true;
-                }
-            } else {    // all types with interger values
-                int value = *((int*)cs->args[j]);
-
-                VariablePool* var = NULL;
-                HASH_FIND_INT(variable_pools[type], &value, var);
-
-                if(var == NULL) {
-                    var = malloc(sizeof(VariablePool));
-                    var->value = value;
-                    var->id = variable_current_ids[type]++;
-                    HASH_ADD_INT(variable_pools[type], value, var);
-
-                    if(dir == DIRECTION_IN)
-                        fprintf(fout, "%s %s_%d = %d;\n", TYPE_STR[type], TYPE_VAR_STR[type], var->id, value);
-                    else
-                        fprintf(fout, "%s %s_%d;\n", TYPE_STR[type], TYPE_VAR_STR[type], var->id);
-                }
-            }
-        }
-    }
-
-    fclose(fout);
-}
-
-
-int main(int argc, char** argv) {
-    char *directory = argv[1];
-
-    printf("Global Metadata\n");
-    GlobalMetadata gm;
-    read_global_metadata(directory, &gm);
-
-    /*
-    printf("Local Metadata\n");
-    for(int i = 0; i < gm.ranks; i++) {
-        LocalMetadata lm;
-        read_local_metadata(path, i, &lm);
-    }
-    */
-
-    int num_sigs;
-    CallSignature *call_sigs = read_cst(directory, &num_sigs);
-    write_init_variables(directory, call_sigs, num_sigs);
-    read_cfg(directory, gm.ranks, call_sigs);
-
-    for(int i = 0; i < num_sigs; i++) {
-        for(int j = 0; j < call_sigs[i].arg_count; j++)
-            free(call_sigs[i].args[j]);
-        free(call_sigs[i].arg_sizes);
-        free(call_sigs[i].arg_types);
-        free(call_sigs[i].arg_directions);
-        free(call_sigs[i].args);
-    }
-    free(call_sigs);
-
-
-    return 0;
-}
